% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvcrand_package.R
\docType{package}
\name{cvcrand}
\alias{cvcrand}
\alias{cvcrand-package}
\title{cvcrand: a package for efficient design and analysis of cluster randomized trials}
\description{
cvcrand: A Package for Covariate-constrained Randomization and the Clustered Permutation Test for
Cluster Randomized Trials
}
\section{cvcrand functions}{

The cvrall function performs constrained randomization on cluster randomized trials (CRTs)
The cvrcov function performs covariate-constrained randomization on cluster randomized trials (CRTs)
The cptest function performs permutation test on the outcome of cluster randomized trials (CRTs)
}

\author{
\strong{Maintainer}: Hengshi Yu \email{hengshi@umich.edu}

Authors:
\itemize{
  \item Fan Li \email{fan.f.li@yale.edu}
  \item John A. Gallis \email{john.gallis@duke.edu}
  \item Elizabeth L. Turner \email{liz.turner@duke.edu}
}

}
\keyword{internal}
